# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CineConnect API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Profile not found
    put:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List/search users
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: skill
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
  /users/{userId}:
    get:
      operationId: getUserById
      tags: [users]
      summary: Get user profile by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
  /users/{userId}/view:
    post:
      operationId: recordProfileView
      tags: [users]
      summary: Record a profile view
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: View recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required: [success]
  /users/{userId}/block:
    post:
      operationId: blockUser
      tags: [users]
      summary: Block a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User blocked
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required: [success]
...
components:
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        clerkId:
          type: string
        name:
          type: string
        email:
          type: string
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
          description: "Actor | Director | Cameraman | Editor | Music | Writer | Producer | Other"
        bio:
          type: ["string", "null"]
        skills:
          type: array
          items:
            type: string
        location:
          type: ["string", "null"]
        portfolioUrl:
          type: ["string", "null"]
        portfolioImages:
          type: array
          items:
            type: string
        isVerified:
          type: boolean
        onlineStatus:
          type: string
          description: "online | offline | away"
        averageRating:
          type: ["number", "null"]
        totalReviews:
          type: integer
        totalProjectsPosted:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - clerkId
        - name
        - email
        - role
        - skills
        - portfolioImages
        - isVerified
        - onlineStatus
        - totalReviews
        - totalProjectsPosted
        - createdAt
    UpdateProfileBody:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        bio:
          type: string
        skills:
          type: array
          items:
            type: string
        location:
          type: string
        portfolioUrl:
          type: string
        portfolioImages:
          type: array
          items:
            type: string
        avatarUrl:
          type: string
        onlineStatus:
          type: string
    UserListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/UserProfile"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - users
        - total
        - page
        - limit
...